11. modify user_id by todo destroyCompleted

+++ app/Http/Controllers/TodosController.php
@@ -91,11 +91,29 @@ class TodosController extends Controller
 
     public function destroyCompleted(Request $request)
     {
+        // [6,9] todo ids we are passing in and want to delete
+        // [5,6,9] all of the users todo ids
+
+        $todosToDelete = $request->todos;
+
+        $userTodoIds = auth()->user()->todos->map(function ($todo) {
+            return $todo->id;
+        });
+
+        $valid = collect($todosToDelete)->every(function ($value, $key) use ($userTodoIds) {
+            return $userTodoIds->contains($value);
+        });
+
+        if (!$valid) {
+            return response()->json('Unauthorized', 401);
+        }
+
         $request->validate([
-            'todos'=>'required|array'
+            'todos' => 'required|array',
         ]);
+
         Todo::destroy($request->todos);
 
-        return response($request->todos,200);
+        return response()->json('Deleted', 200);
     }
 }
